refactor(dashboard): extract ProfileContactSidebar bento + header sections
Behavior-preserving move (ProfileContactSidebar.tsx 1042 -> 794): the two
prop-driven presentation sections (ProfileBentoView,
ProfileSidebarHeaderCard) plus their block-only helper toPreviewSocialLinks
move to ProfileContactSidebarSections.tsx (same dir, boundaries-safe).
Only the 2 components are exported; toPreviewSocialLinks (0 uses in the
main body) travels with them. The shared tab/category helpers stay in
ProfileContactSidebar.tsx. Bodies verbatim.
